adjust styles
