+++ src/components/layout/index.tsx
@@ -2,8 +2,7 @@ import React, { useContext } from 'react';
 import { useRouter } from 'next/router';
 import { NextSeo } from 'next-seo';
 import validator from 'validator';
-import Head from 'next/head';
-import classnames from 'classnames';
+import Head from 'next/head';import classnames from 'classnames';
 import { useTranslation } from 'i18n';
 import {
   MobileNav,
