+++ src/screens/home/components/index.ts
@@ -1,5 +1,7 @@
 import LatestBlocks from './latest_blocks';
+import StabilitiesItem from './stabilities';
 
 export {
   LatestBlocks,
+  StabilitiesItem,
 };
