+++ src/screens/home/styles.tsx
@@ -19,8 +19,8 @@ export const useGetStyles = () => {
         background: 'pink',
       },
       '& .stablities': {
-        background: theme?.palette?.background?.paper,
-        borderRadius: '2px',
+        // background: theme?.palette?.background?.paper,
+        // borderRadius: '2px',
       },
       '& .consensus': {
         background: 'pink',
